+++ server/exceptions.py
@@ -16,3 +16,4 @@ class ConfigError(TaskException):
 class ProcessException(Exception):
     def __init__(self, message):
         super().__init__(message)
+
